requirements.txt
customtkinter==5.2.0
darkdetect==0.8.0
packaging==23.1
Pillow==10.0.0
pytesseract==0.3.10
